pharma_parser/requirements.txt
# Django
django==4.2.9
djangorestframework==3.14.0
django-environ==0.11.2

# Celery
celery==5.3.6
redis==5.0.1

# PDF processing
pdfplumber==0.11.0
pdf2image==1.17.0
Pillow==10.2.0
PyMuPDF==1.24.3

# OCR
surya-ocr==0.4.15

# LangChain + LLM
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.32

# Data
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1

# File watching (reuse existing)
watchdog==4.0.0
